+++ R/ggraph-package.R
@@ -8,9 +8,5 @@
 #' @import ggplot2 tidygraph
 #' @importFrom Rcpp sourceCpp
 #' @importFrom rlang .data
-#' @rawNamespace
-#' if (getRversion() >= "3.6.0") {
-#'   S3method(gganimate::layer_type, GeomEdgePath)
-#' }
 ## usethis namespace: end
 NULL